Found final bug
Was able to get the project to work by switching out RA2 input for RB9, it seems that RB1 and RA2 were not being read correctly, causing the LED to stay illuminated at all times.

I will send out an email to Rushi  this evening/ in the morning, there may have been something he overlooked on the PIC  before he created the assignment.